backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
python-multipart
pytesseract
Pillow
requests
argon2-cffi
email-validator
google-generativeai

